refactor: clarify command architecture
