+++ src/client-command-registry.ts
@@ -1,42 +1,4 @@
-export const CLIENT_COMMANDS = {
-  alert: 'alert',
-  appState: 'appstate',
-  appSwitcher: 'app-switcher',
-  apps: 'apps',
-  back: 'back',
-  batch: 'batch',
-  boot: 'boot',
-  click: 'click',
-  clipboard: 'clipboard',
-  devices: 'devices',
-  diff: 'diff',
-  fill: 'fill',
-  find: 'find',
-  focus: 'focus',
-  get: 'get',
-  home: 'home',
-  is: 'is',
-  keyboard: 'keyboard',
-  logs: 'logs',
-  longPress: 'longpress',
-  network: 'network',
-  perf: 'perf',
-  pinch: 'pinch',
-  press: 'press',
-  push: 'push',
-  record: 'record',
-  replay: 'replay',
-  rotate: 'rotate',
-  scroll: 'scroll',
-  screenshot: 'screenshot',
-  settings: 'settings',
-  snapshot: 'snapshot',
-  swipe: 'swipe',
-  test: 'test',
-  trace: 'trace',
-  triggerAppEvent: 'trigger-app-event',
-  type: 'type',
-  wait: 'wait',
-} as const;
+import { PUBLIC_COMMANDS } from './command-catalog.ts';
 
+export const CLIENT_COMMANDS = PUBLIC_COMMANDS;
 export type ClientCommandName = (typeof CLIENT_COMMANDS)[keyof typeof CLIENT_COMMANDS];
