+++ src/client.ts
@@ -2,6 +2,14 @@ import { sendToDaemon } from './daemon-client.ts';
 import { prepareMetroRuntime, reloadMetro } from './client-metro.ts';
 import { CLIENT_COMMANDS } from './client-command-registry.ts';
 import { createAgentDeviceCommandClient, type PreparedClientCommand } from './client-commands.ts';
+import {
+  elementTargetCodec,
+  fillCommandCodec,
+  findCommandCodec,
+  interactionTargetCodec,
+  isCommandCodec,
+  settingsCommandCodec,
+} from './command-codecs.ts';
 import { throwDaemonError } from './daemon-error.ts';
 import {
   buildFlags,
@@ -36,10 +44,7 @@ import type {
   CaptureSnapshotOptions,
   CaptureSnapshotResult,
   CommandRequestResult,
-  ElementTarget,
   EnsureSimulatorOptions,
-  FindOptions,
-  InteractionTarget,
   InternalRequestOptions,
   Lease,
   MaterializationReleaseOptions,
@@ -342,12 +347,16 @@ export function createAgentDeviceClient(
     },
     interactions: {
       click: async (options) =>
-        await executeCommandRequest(CLIENT_COMMANDS.click, targetPositionals(options), {
+        await executeCommandRequest(CLIENT_COMMANDS.click, interactionTargetCodec.encode(options), {
           ...options,
           clickButton: options.button,
         }),
       press: async (options) =>
-        await executeCommandRequest(CLIENT_COMMANDS.press, targetPositionals(options), options),
+        await executeCommandRequest(
+          CLIENT_COMMANDS.press,
+          interactionTargetCodec.encode(options),
+          options,
+        ),
       longPress: async (options) =>
         await executeCommandRequest(
           CLIENT_COMMANDS.longPress,
@@ -377,7 +386,7 @@ export function createAgentDeviceClient(
       fill: async (options) =>
         await executeCommandRequest(
           CLIENT_COMMANDS.fill,
-          [...targetPositionals(options), options.text],
+          fillCommandCodec.encode(options),
           options,
         ),
       scroll: async (options) =>
@@ -395,21 +404,13 @@ export function createAgentDeviceClient(
       get: async (options) =>
         await executeCommandRequest(
           CLIENT_COMMANDS.get,
-          [options.format, ...elementPositionals(options)],
+          [options.format, ...elementTargetCodec.encode(options)],
           options,
         ),
       is: async (options) =>
-        await executeCommandRequest(
-          CLIENT_COMMANDS.is,
-          [
-            options.predicate,
-            options.selector,
-            ...(options.predicate === 'text' ? [options.value] : []),
-          ],
-          options,
-        ),
+        await executeCommandRequest(CLIENT_COMMANDS.is, isCommandCodec.encode(options), options),
       find: async (options) =>
-        await executeCommandRequest(CLIENT_COMMANDS.find, findPositionals(options), {
+        await executeCommandRequest(CLIENT_COMMANDS.find, findCommandCodec.encode(options), {
           ...options,
           findFirst: options.first,
           findLast: options.last,
@@ -468,28 +469,11 @@ export function createAgentDeviceClient(
       update: async (options) =>
         await executeCommandRequest(
           CLIENT_COMMANDS.settings,
-          [
-            options.setting,
-            options.state,
-            ...('latitude' in options ? [String(options.latitude), String(options.longitude)] : []),
-            ...('permission' in options ? [options.permission] : []),
-            ...('mode' in options && options.mode ? [options.mode] : []),
-          ],
+          settingsCommandCodec.encode(options),
           options,
         ),
     },
   };
-}
-
-function targetPositionals(options: InteractionTarget): string[] {
-  if (options.ref !== undefined) return [options.ref, ...optionalString(options.label)];
-  if (options.selector !== undefined) return [options.selector];
-  return [String(options.x), String(options.y)];
-}
-
-function elementPositionals(options: ElementTarget): string[] {
-  if (options.ref !== undefined) return [options.ref, ...optionalString(options.label)];
-  return [options.selector];
 }
 
 function stringifyPayload(payload: AppPushOptions['payload']): string {
@@ -498,29 +482,6 @@ function stringifyPayload(payload: AppPushOptions['payload']): string {
 
 function triggerEventPositionals(options: AppTriggerEventOptions): string[] {
   return [options.event, ...(options.payload ? [JSON.stringify(options.payload)] : [])];
-}
-
-function findPositionals(options: FindOptions): string[] {
-  const args =
-    options.locator && options.locator !== 'any'
-      ? [options.locator, options.query]
-      : [options.query];
-  switch (options.action) {
-    case undefined:
-    case 'click':
-    case 'focus':
-    case 'exists':
-      return options.action ? [...args, options.action] : args;
-    case 'getText':
-      return [...args, 'get', 'text'];
-    case 'getAttrs':
-      return [...args, 'get', 'attrs'];
-    case 'wait':
-      return [...args, 'wait', ...optionalNumber(options.timeoutMs)];
-    case 'fill':
-    case 'type':
-      return [...args, options.action, options.value];
-  }
 }
 
 function logsPositionals(options: { action?: string; message?: string }): string[] {
@@ -575,91 +536,4 @@ function normalizeLease(data: Record<string, unknown>): Lease {
   };
 }
 
-export type {
-  AgentDeviceClient,
-  AgentDeviceClientConfig,
-  AgentDeviceCommandClient,
-  AgentDeviceDaemonTransport,
-  AgentDeviceDevice,
-  AgentDeviceIdentifiers,
-  AgentDeviceRequestOverrides,
-  AgentDeviceSelectionOptions,
-  AgentDeviceSession,
-  AgentDeviceSessionDevice,
-  AlertCommandOptions,
-  AlertCommandResult,
-  AppPushOptions,
-  AppStateCommandOptions,
-  AppStateCommandResult,
-  AppTriggerEventOptions,
-  AppCloseOptions,
-  AppCloseResult,
-  AppDeployOptions,
-  AppDeployResult,
-  AppInstallFromSourceOptions,
-  AppInstallFromSourceResult,
-  AppListOptions,
-  AppOpenOptions,
-  AppOpenResult,
-  AppSwitcherCommandOptions,
-  AppSwitcherCommandResult,
-  BackCommandOptions,
-  BackCommandResult,
-  CaptureScreenshotOptions,
-  CaptureScreenshotResult,
-  CaptureSnapshotOptions,
-  CaptureSnapshotResult,
-  CaptureDiffOptions,
-  ClipboardCommandOptions,
-  ClipboardCommandResult,
-  CommandRequestResult,
-  BatchRunOptions,
-  BatchStep,
-  ClickOptions,
-  ElementTarget,
-  DeviceBootOptions,
-  EnsureSimulatorOptions,
-  EnsureSimulatorResult,
-  FillOptions,
-  FindLocator,
-  FindOptions,
-  FocusOptions,
-  GetOptions,
-  HomeCommandOptions,
-  HomeCommandResult,
-  IsOptions,
-  InteractionTarget,
-  KeyboardCommandOptions,
-  KeyboardCommandResult,
-  Lease,
-  LeaseAllocateOptions,
-  LeaseOptions,
-  LeaseScopedOptions,
-  LogsOptions,
-  LongPressOptions,
-  MaterializationReleaseOptions,
-  MaterializationReleaseResult,
-  MetroPrepareOptions,
-  MetroPrepareResult,
-  MetroReloadOptions,
-  MetroReloadResult,
-  NetworkOptions,
-  PerfOptions,
-  PermissionTarget,
-  PinchOptions,
-  PressOptions,
-  RecordOptions,
-  ReplayRunOptions,
-  ReplayTestOptions,
-  RotateCommandOptions,
-  RotateCommandResult,
-  ScrollOptions,
-  SessionCloseResult,
-  SettingsUpdateOptions,
-  StartupPerfSample,
-  SwipeOptions,
-  TraceOptions,
-  TypeTextOptions,
-  WaitCommandOptions,
-  WaitCommandResult,
-} from './client-types.ts';
+export type * from './client-types.ts';
