+++ src/command-catalog.ts
@@ -1,0 +1,88 @@
+export const PUBLIC_COMMANDS = {
+  alert: 'alert',
+  appState: 'appstate',
+  appSwitcher: 'app-switcher',
+  apps: 'apps',
+  back: 'back',
+  batch: 'batch',
+  boot: 'boot',
+  click: 'click',
+  clipboard: 'clipboard',
+  devices: 'devices',
+  diff: 'diff',
+  fill: 'fill',
+  find: 'find',
+  focus: 'focus',
+  get: 'get',
+  home: 'home',
+  is: 'is',
+  keyboard: 'keyboard',
+  logs: 'logs',
+  longPress: 'longpress',
+  network: 'network',
+  perf: 'perf',
+  pinch: 'pinch',
+  press: 'press',
+  push: 'push',
+  record: 'record',
+  replay: 'replay',
+  rotate: 'rotate',
+  scroll: 'scroll',
+  screenshot: 'screenshot',
+  settings: 'settings',
+  snapshot: 'snapshot',
+  swipe: 'swipe',
+  test: 'test',
+  trace: 'trace',
+  triggerAppEvent: 'trigger-app-event',
+  type: 'type',
+  wait: 'wait',
+} as const;
+
+export const INTERNAL_COMMANDS = {
+  ensureSimulator: 'ensure-simulator',
+  installSource: 'install_source',
+  leaseAllocate: 'lease_allocate',
+  leaseHeartbeat: 'lease_heartbeat',
+  leaseRelease: 'lease_release',
+  releaseMaterializedPaths: 'release_materialized_paths',
+  sessionList: 'session_list',
+} as const;
+
+export const DAEMON_COMMAND_GROUPS = {
+  inventory: commandSet(
+    INTERNAL_COMMANDS.sessionList,
+    INTERNAL_COMMANDS.ensureSimulator,
+    PUBLIC_COMMANDS.devices,
+    PUBLIC_COMMANDS.apps,
+  ),
+  state: commandSet(PUBLIC_COMMANDS.boot, PUBLIC_COMMANDS.appState),
+  observability: commandSet(PUBLIC_COMMANDS.perf, PUBLIC_COMMANDS.logs, PUBLIC_COMMANDS.network),
+  replay: commandSet(PUBLIC_COMMANDS.replay, PUBLIC_COMMANDS.test),
+  snapshot: commandSet(
+    PUBLIC_COMMANDS.snapshot,
+    PUBLIC_COMMANDS.diff,
+    PUBLIC_COMMANDS.wait,
+    PUBLIC_COMMANDS.alert,
+    PUBLIC_COMMANDS.settings,
+  ),
+  selectorValidationExempt: commandSet(
+    INTERNAL_COMMANDS.sessionList,
+    PUBLIC_COMMANDS.devices,
+    INTERNAL_COMMANDS.ensureSimulator,
+    INTERNAL_COMMANDS.releaseMaterializedPaths,
+  ),
+  leaseAdmissionExempt: commandSet(
+    INTERNAL_COMMANDS.sessionList,
+    PUBLIC_COMMANDS.devices,
+    INTERNAL_COMMANDS.ensureSimulator,
+    INTERNAL_COMMANDS.releaseMaterializedPaths,
+    INTERNAL_COMMANDS.leaseAllocate,
+    INTERNAL_COMMANDS.leaseHeartbeat,
+    INTERNAL_COMMANDS.leaseRelease,
+  ),
+} as const;
+
+function commandSet(...commands: readonly string[]): ReadonlySet<string> {
+  return new Set(commands);
+}
